add main class to write on the file

+++ src/main/java/edu/ppu/services/SumOverMaxService.java
@@ -13,7 +13,7 @@ public class SumOverMaxService {
         this.listSumService = listSumService;
     }
 
-    public float getSumOverMax(List<Integer> numbers) {
+    public int getSumOverMax(List<Integer> numbers) {
 
         if (numbers == null || numbers.size() == 0) {
             throw new RuntimeException("list is empty");
